feat: Add manual refresh capability to HomeScreen
- Added an IconButton next to the location name to allow manual weather updates.
- Integrated ref.read() call to trigger the fetchWeather() method from weatherProvider.
- Improved UX by providing a direct way to sync data without restarting the app.

+++ lib/widgets/hours_card_widget.dart
@@ -47,7 +47,7 @@ class HoursCardWidget extends ConsumerWidget {
 
           // Kelvin temperature ko Celsius mein badal kar round kiya
           String temp =
-              ((hoursWeather.hourlyWeather[index]['temp'] as double) -
+              ((hoursWeather.hourlyWeather[index]['temp']) -
                       273.15)
                   .round()
                   .toString();
